+++ internal/model/aff_campaign.go
@@ -118,7 +118,7 @@ func (m *AffCampaign) ToDto() dto.AffCampaignDto {
 
 type AffCampaignApp struct {
 	ID                uint64                 `gorm:"primarykey" json:"id"`
-	BrandId           uint64                 `json:"brand_id"`
+	BrandId           uint                   `json:"brand_id"`
 	Brand             Brand                  `json:"brand" gorm:"foreignKey:BrandId"`
 	AccessTradeId     string                 `json:"accesstrade_id" gorm:"column:accesstrade_id"`
 	CreatedAt         time.Time              `json:"created_at"`
@@ -167,7 +167,7 @@ func (m *AffCampaignApp) ToAffCampaignAppDto() dto.AffCampaignAppDto {
 type AffCampaignComBrand struct {
 	ID           uint64                 `gorm:"primarykey" json:"id"`
 	Name         string                 `json:"name"`
-	BrandId      uint64                 `json:"brand_id"`
+	BrandId      uint                   `json:"brand_id"`
 	Brand        Brand                  `json:"brand" gorm:"foreignKey:BrandId"`
 	StellaMaxCom string                 `json:"stella_max_com"`
 	Attributes   []AffCampaignAttribute `json:"attributes" gorm:"foreignKey:CampaignId"`
@@ -195,7 +195,7 @@ func (m *AffCampaignComBrand) ToAffCampaignLessDto() dto.AffCampaignLessDto {
 type AffCampComFavBrand struct {
 	ID            uint64                 `gorm:"primarykey" json:"id"`
 	Name          string                 `json:"name"`
-	BrandId       uint64                 `json:"brand_id"`
+	BrandId       uint                   `json:"brand_id"`
 	FavoriteBrand UserFavoriteBrand      `gorm:"foreignKey:BrandId;references:BrandId" json:"favrorite_brand"`
 	StellaMaxCom  string                 `json:"stella_max_com"`
 	Attributes    []AffCampaignAttribute `json:"attributes" gorm:"foreignKey:CampaignId"`
